% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_linreg.R
\name{mcmc_linreg}
\alias{mcmc_linreg}
\title{MCMC for Bayesian Linear Regression}
\usage{
mcmc_linreg(y, X, beta_init, sigma2, tau2, n_iter, proposal_sd)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{X}{A numeric matrix of predictors (n x p).}

\item{beta_init}{Initial values for the regression coefficients.}

\item{sigma2}{Known error variance.}

\item{tau2}{Prior variance for the coefficients.}

\item{n_iter}{Number of MCMC iterations.}

\item{proposal_sd}{Standard deviation of the proposal distribution for Metropolis steps.}
}
\value{
A matrix of size n_iter x p containing posterior samples of the coefficients.
}
\description{
This function is implemented in C++ and exposed to R via Rcpp. It performs MCMC sampling 
for Bayesian linear regression using a Metropolis-Hastings algorithm.
}
