\name{SA24204126-package}
\alias{SA24204126-package}
\alias{SA24204126}
\docType{package}
\title{
  Bayesian Linear Regression with MCMC
}
\description{
  `SA24204126`包提供了一个简洁的框架来进行**贝叶斯线性回归**分析，包括：  
- 数据模拟函数 `simulate_data()`  
- 基于Metropolis-Hastings算法的后验抽样函数 `mcmc_linreg()`  
- 对后验结果进行汇总的函数 `summarize_posterior()`
}
\details{
  null.
}
\author{
YanBookStore
Maintainer: YanBookStore <2420253242@qq.com>
}
\keyword{ package }

